+++ pkg/protocol/decoder/influxdb/decoder.go
@@ -15,6 +15,7 @@
 package influxdb
 
 import (
+	"fmt"
 	"net/http"
 	"strconv"
 	"strings"
@@ -46,13 +47,41 @@ const (
 
 const tagDB = "__tag__:db"
 
+const metadataKeyDB = "db"
+
+const (
+	formDataKeyDB        = "db"
+	formDataKeyPrecision = "precision"
+)
+
 // Decoder impl
 type Decoder struct {
 	FieldsExtend bool
 }
 
 func (d *Decoder) Decode(data []byte, req *http.Request, tags map[string]string) (logs []*protocol.Log, decodeErr error) {
-	precision := req.FormValue("precision")
+	points, err := d.decodeToInfluxdbPoints(data, req)
+	if err != nil {
+		return nil, err
+	}
+	logs = d.parsePointsToLogs(points, req)
+	return logs, nil
+}
+
+func (d *Decoder) ParseRequest(res http.ResponseWriter, req *http.Request, maxBodySize int64) (data []byte, statusCode int, err error) {
+	return common.CollectBody(res, req, maxBodySize)
+}
+
+func (d *Decoder) DecodeV2(data []byte, req *http.Request) ([]*imodels.PipelineGroupEvents, error) {
+	points, err := d.decodeToInfluxdbPoints(data, req)
+	if err != nil {
+		return nil, err
+	}
+	return d.parsePointsToEvents(points, req)
+}
+
+func (d *Decoder) decodeToInfluxdbPoints(data []byte, req *http.Request) ([]models.Point, error) {
+	precision := req.FormValue(formDataKeyPrecision)
 	var points []models.Point
 	var err error
 	if precision != "" {
@@ -63,19 +92,72 @@ func (d *Decoder) Decode(data []byte, req *http.Request, tags map[string]string)
 	if err != nil {
 		return nil, err
 	}
-
-	logs = d.parsePointsToLogs(points, req)
-
-	return logs, err
+	return points, nil
 }
 
-func (d *Decoder) ParseRequest(res http.ResponseWriter, req *http.Request, maxBodySize int64) (data []byte, statusCode int, err error) {
-	return common.CollectBody(res, req, maxBodySize)
+func (d *Decoder) parsePointsToEvents(points []models.Point, req *http.Request) ([]*imodels.PipelineGroupEvents, error) {
+	group := &imodels.PipelineGroupEvents{
+		Group:  imodels.NewGroup(imodels.NewMetadata(), imodels.NewTags()),
+		Events: make([]imodels.PipelineEvent, 0, len(points)),
+	}
+	if db := req.FormValue(formDataKeyDB); len(db) > 0 {
+		group.Group.Metadata.Add(metadataKeyDB, db)
+	}
+	for _, point := range points {
+		tags := imodels.NewTags()
+		point.ForEachTag(func(k, v []byte) bool {
+			tags.Add(string(k), string(v)) // deep copy here
+			return true
+		})
+		multiValue, typedValue, err := d.parsePointFieldsToMetricValues(point)
+		if err != nil {
+			return nil, err
+		}
+		metric := imodels.NewMetric(string(point.Name()), imodels.MetricTypeUntyped, tags, point.UnixNano(), multiValue, typedValue)
+		group.Events = append(group.Events, metric)
+	}
+	return []*imodels.PipelineGroupEvents{group}, nil
 }
 
-func (d *Decoder) DecodeV2(data []byte, req *http.Request) (groups []*imodels.PipelineGroupEvents, err error) {
-	//TODO: Implement DecodeV2
-	return nil, nil
+func (d *Decoder) parsePointFieldsToMetricValues(p models.Point) (imodels.MetricValue, imodels.MetricTypedValues, error) {
+	multiValue := imodels.NewMetricMultiValue()
+	typedValue := imodels.NewMetricTypedValues()
+
+	iter := p.FieldIterator()
+	for iter.Next() {
+		if len(iter.FieldKey()) == 0 {
+			continue
+		}
+		switch iter.Type() {
+		case models.Float:
+			v, err := iter.FloatValue()
+			if err != nil {
+				return nil, nil, fmt.Errorf("unable to unmarshal field %s: %s", string(iter.FieldKey()), err)
+			}
+			multiValue.Add(string(iter.FieldKey()), v)
+		case models.Integer:
+			v, err := iter.IntegerValue()
+			if err != nil {
+				return nil, nil, fmt.Errorf("unable to unmarshal field %s: %s", string(iter.FieldKey()), err)
+			}
+			multiValue.Add(string(iter.FieldKey()), float64(v))
+		case models.Unsigned:
+			v, err := iter.UnsignedValue()
+			if err != nil {
+				return nil, nil, fmt.Errorf("unable to unmarshal field %s: %s", string(iter.FieldKey()), err)
+			}
+			multiValue.Add(string(iter.FieldKey()), float64(v))
+		case models.String:
+			typedValue.Add(string(iter.FieldKey()), &imodels.TypedValue{Type: imodels.ValueTypeString, Value: iter.StringValue()})
+		case models.Boolean:
+			v, err := iter.BooleanValue()
+			if err != nil {
+				return nil, nil, fmt.Errorf("unable to unmarshal field %s: %s", string(iter.FieldKey()), err)
+			}
+			typedValue.Add(string(iter.FieldKey()), &imodels.TypedValue{Type: imodels.ValueTypeBoolean, Value: v})
+		}
+	}
+	return multiValue, typedValue, nil
 }
 
 func (d *Decoder) parsePointsToLogs(points []models.Point, req *http.Request) []*protocol.Log {