+++ plugins/input/opentelemetry/service_otlp_test.go
@@ -555,11 +555,11 @@ func initExponentialHistogramMetric(hm pmetric.Metric) {
 	hdp0.Negative().BucketCounts().FromRaw([]uint64{1, 1})
 
 	// The above will print:
-	// Bucket (-1.414214, -1.000000], Count: 1
-	// Bucket (-1.000000, -0.707107], Count: 1
+	// Bucket (2.000000, 2.828427], Count: 1
+	// Bucket (1.414214, 2.000000], Count: 1
 	// Bucket [0, 0], Count: 1
-	// Bucket [0.707107, 1.000000), Count: 1
-	// Bucket [1.000000, 1.414214), Count: 1
+	// Bucket [-1.000000, -0.707107), Count: 1
+	// Bucket [-1.414214, -1.000000), Count: 1
 
 	hdp1 := hdps.AppendEmpty()
 	initMetricAttributes2(hdp1.Attributes())
@@ -578,8 +578,8 @@ func initExponentialHistogramMetric(hm pmetric.Metric) {
 
 	// The above will print:
 	// Bucket [0, 0], Count: 1
-	// Bucket [0.250000, 1.000000), Count: 1
-	// Bucket [1.000000, 4.000000), Count: 1
+	// Bucket (0.250000, 1.000000], Count: 1
+	// Bucket (1.000000, 4.000000], Count: 1
 
 	exemplar := hdp1.Exemplars().AppendEmpty()
 	exemplar.SetTimestamp(metricExemplarTimestamp)